refactor chart explanation popover to reduce code duplication
